PP-1059 Refactor audiobooks (#425)
* Update Audiobook classes

* building open access happy path

* Launch audiobook from delegate extension

* Merge hotfix into main version 1.0.37 (#414)

* Fix OOM issue appearing on iPhone XR

This could potentially be an issue on other devices as well, since
we are trying to load a large binary into memory to try to parse
it as json. In the case of LCP this isn't what we want to do.

* Rebase off of main (#411)

* Update check-build-number.yml (#410)

What's this do?
Updates workflows

Why are we doing this? (w/ Notion link if applicable)
To support pushing TestFlight builds for merges into the hotfix branch

* Update upload-on-merge.yml (#409)

* Update build #

* Update project.pbxproj (#413)

Bump marketing version to create build

* Update project.pbxproj

---------


* Update project.pbxproj (#415)

Revert version and build numbers

* Update TPPBookCellDelegate+Extensions.swift

* Integrate updated audiobook changes

* Update check-build-number.yml

* Update upload-on-merge.yml

* Clean up

* Update .gitmodules

* Update project.pbxproj

* Update project.pbxproj

* Update project.pbxproj

* Update project.pbxproj

* Update project.pbxproj

* Delete ios-drm-audioengine

* Update scripts

* Update ios-audiobooktoolkit

* Update AudiobookBookmarkBusinessLogicTests.swift

* [PP-1059] Integrates refactor of Audiobook Library (#416)

* Update Audiobook classes

* building open access happy path

* Launch audiobook from delegate extension

* Update TPPBookCellDelegate+Extensions.swift

* Integrate updated audiobook changes

* Clean up

* Update .gitmodules

* Update project.pbxproj

* Update project.pbxproj

* Update project.pbxproj

* Update project.pbxproj

* Update project.pbxproj

* Delete ios-drm-audioengine

* Update scripts

* Update ios-audiobooktoolkit

* Update AudiobookBookmarkBusinessLogicTests.swift

* Update project.pbxproj

* Revert "Update project.pbxproj"

This reverts commit abe6dc13690b2383ea52d027bb2c0378c9c5ab1a.

* Update project.pbxproj

* Update project.pbxproj

* Revert "Update project.pbxproj"

This reverts commit 9aca7790ed506824cd474317c3f079d452d1f179.

* Restore Build (#417)

* Update Audiobook classes

* building open access happy path

* Launch audiobook from delegate extension

* Update TPPBookCellDelegate+Extensions.swift

* Integrate updated audiobook changes

* Clean up

* Update .gitmodules

* Update project.pbxproj

* Update project.pbxproj

* Update project.pbxproj

* Update project.pbxproj

* Update project.pbxproj

* Delete ios-drm-audioengine

* Update scripts

* Update ios-audiobooktoolkit

* Update AudiobookBookmarkBusinessLogicTests.swift

* Update project.pbxproj

* Revert "Update project.pbxproj"

This reverts commit abe6dc13690b2383ea52d027bb2c0378c9c5ab1a.

* Update project.pbxproj

* Update project.pbxproj

* Revert "Update project.pbxproj"

This reverts commit 9aca7790ed506824cd474317c3f079d452d1f179.

* Update ios-audiobook-overdrive

* Reapply "Update project.pbxproj"

This reverts commit 7ed72dded4b5fc59496a290aa82f2c32fb0cb2f6.

* Revert "Restore Build (#417)"

This reverts commit 933de16e4187c028ff64cfc385342de02dadf048.

* Update ios-audiobook-overdrive (#419)

* Fix build (#420)

* Update ios-audiobook-overdrive

* Update project.pbxproj

* Fix build (#421)

* Update ios-audiobook-overdrive

* Update project.pbxproj

* Update ios-audiobooktoolkit

* Fix build  (#422)

* Update ios-audiobook-overdrive

* Update project.pbxproj

* Update ios-audiobooktoolkit

* Update project.pbxproj

* Update project.pbxproj

---------

+++ Palace/MyBooks/MyBooksDownloadCenter.swift
@@ -433,67 +433,72 @@ extension MyBooksDownloadCenter {
     
     do {
       switch book.defaultBookContentType {
-      case .epub:
-        try FileManager.default.removeItem(at: bookURL)
-      case .audiobook:
-        deleteLocalContent(forAudiobook: book, at: bookURL)
-      case .pdf:
+      case .epub, .pdf:
         try FileManager.default.removeItem(at: bookURL)
 #if LCP
-        try LCPPDFs.deletePdfContent(url: bookURL)
+        if book.defaultBookContentType == .pdf {
+          try LCPPDFs.deletePdfContent(url: bookURL)
+        }
 #endif
+      case .audiobook:
+        deleteLocalContent(forAudiobook: book, at: bookURL)
       case .unsupported:
-        break
+        NSLog("Unsupported content type for deletion.")
       }
     } catch {
-      NSLog("Failed to remove local content for download: \(error.localizedDescription)")
+      NSLog("Failed to remove local content for book with identifier \(identifier): \(error.localizedDescription)")
     }
   }
-  
+
   func deleteLocalContent(forAudiobook book: TPPBook, at bookURL: URL) {
-    guard let data = try? Data(contentsOf: bookURL) else { return }
-
-    var dict = [String: Any]()
-    
 #if FEATURE_OVERDRIVE
-    if book.distributor == OverdriveDistributorKey {
-      if let json = try? JSONSerialization.jsonObject(with: data, options: []) as? [String: Any] {
-        dict = json ?? [String: Any]()
-      }
-
-      dict["id"] = book.identifier
+    if book.distributor == OverdriveDistributorKey,
+        let data = try? Data(contentsOf: bookURL),
+        let json = try? JSONSerialization.jsonObject(with: data, options: []),
+        let jsonDict = json as? [String: Any] {
+      handleOverdriveContentDeletion(jsonDict, for: book)
     }
 #endif
     
 #if LCP
     if LCPAudiobooks.canOpenBook(book) {
-      let lcpAudiobooks = LCPAudiobooks(for: bookURL)
-      lcpAudiobooks?.contentDictionary { dict, error in
-        if let _ = error {
-          // LCPAudiobooks logs this error
-          return
-        }
-        if let dict = dict {
-          // Delete decrypted content for the book
-          let mutableDict = dict.mutableCopy() as? [String: Any]
-          AudiobookFactory.audiobook(mutableDict ?? [:])?.deleteLocalContent()
-        }
-      }
-      // Delete LCP book file
-      if FileManager.default.fileExists(atPath: bookURL.path) {
-        do {
-          try FileManager.default.removeItem(at: bookURL)
-        } catch {
-          TPPErrorLogger.logError(error, summary: "Failed to delete LCP audiobook local content", metadata: ["book": book.loggableShortString()])
-        }
-      }
+      deleteLCPLocalContent(for: book, at: bookURL)
     } else {
-      // Not an LCP book
-      AudiobookFactory.audiobook(dict)?.deleteLocalContent()
+      deleteNonLCPLocalContent(at: bookURL)
     }
 #else
-    AudiobookFactory.audiobook(dict)?.deleteLocalContent()
+    deleteNonLCPLocalContent(at: bookURL)
 #endif
+  }
+  
+  private func handleOverdriveContentDeletion(_ dict: [String: Any], for book: TPPBook) {
+    // Implement specific deletion logic for Overdrive content if required.
+  }
+  
+#if LCP
+  private func deleteLCPLocalContent(for book: TPPBook, at bookURL: URL) {
+    LCPAudiobooks(for: bookURL)?.contentDictionary { dict, error in
+      guard error == nil, let dict = dict as? [String: Any] else { return }
+      self.deleteAudiobookContent(dict, for: book, bookURL: bookURL)
+    }
+  }
+  
+  private func deleteAudiobookContent(_ dict: [String: Any], for book: TPPBook, bookURL: URL) {
+    // Assuming we create and use an audiobook object to manage deletion
+    guard let jsonData = try? JSONSerialization.data(withJSONObject: dict, options: []),
+          let manifest = try? Manifest.customDecoder().decode(Manifest.self, from: jsonData),
+          let audiobook = AudiobookFactory.audiobook(for: manifest, bookIdentifier: book.identifier, decryptor: nil, token: book.bearerToken) else {
+      return
+    }
+    audiobook.deleteLocalContent(completion: { _, _ in })
+    if FileManager.default.fileExists(atPath: bookURL.path) {
+      try? FileManager.default.removeItem(at: bookURL)
+    }
+  }
+#endif
+  
+  private func deleteNonLCPLocalContent(at bookURL: URL) {
+    try? FileManager.default.removeItem(at: bookURL)
   }
   
   @objc func returnBook(withIdentifier identifier: String, completion: (() -> Void)? = nil) {
@@ -813,7 +818,6 @@ extension MyBooksDownloadCenter: URLSessionTaskDelegate {
         completionHandler(request)
         return
       }
-      
       
       // Prevent redirection from HTTPS to a non-HTTPS URL.
       if task.originalRequest?.url?.scheme == "https" && request.url?.scheme != "https" {
